完成分类模块前后端联调，新增分类树状结构查询方法，优化项目结构

+++ src/main/java/cc/jiusi/blog/service/impl/CategoryServiceImpl.java
@@ -1,8 +1,11 @@
 package cc.jiusi.blog.service.impl;
 
+import cc.jiusi.blog.common.constants.SysConstants;
 import cc.jiusi.blog.common.constants.UserConstants;
 import cc.jiusi.blog.common.res.ResultCodeEnum;
+import cc.jiusi.blog.entity.dto.CategoryDto;
 import cc.jiusi.blog.entity.po.Category;
+import cc.jiusi.blog.entity.po.Tag;
 import cc.jiusi.blog.entity.vo.CategoryVo;
 import cc.jiusi.blog.exception.GlobalException;
 import cc.jiusi.blog.mapper.CategoryMapper;
@@ -17,7 +20,7 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import java.util.*;
 import java.util.stream.Collectors;
 
 /**
@@ -93,6 +96,76 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
     /**
      * @author: 九思.
      * @date: 2023/11/17 20:07
+     * @param: Long> ids
+     * @return: void
+     * @description: 查询分类列表（树结构）
+     */
+    @Override
+    public List<CategoryVo> selectTreeList(CategoryDto dto) {
+        // 构造查询条件
+        LambdaQueryWrapper<Category> wrapper = Wrappers.<Category>lambdaQuery();
+        String name = dto.getName();
+        String status = dto.getStatus();
+        Date[] createTime = dto.getCreateTime();
+        wrapper.like(name != null, Category::getName, name)
+                .eq(status != null, Category::getStatus, status)
+                .orderByAsc(Category::getOrder);
+        // 处理时间数组
+        if (createTime != null) {
+            wrapper.ge(createTime[0] != null, Category::getCreateTime, createTime[0])
+                    .le(createTime[1] != null, Category::getCreateTime, createTime[1]);
+        }
+        List<Category> categories = categoryMapper.selectList(wrapper);
+        // 将 Po 集合 转为 Vo ，向前端展示
+        List<CategoryVo> voList = categories.stream().map(category -> {
+            CategoryVo vo = new CategoryVo();
+            BeanUtils.copyProperties(category, vo);
+            return vo;
+        }).collect(Collectors.toList());
+        // 构造树结构
+        return buildTree(voList);
+    }
+
+    /**
+     * @author: 九思.
+     * @date: 2023/12/11 22:12
+     * @param:  CategoryVo> categoryList
+     * @return: List<CategoryVo>
+     * @description: 构建分类树状结构
+     */
+    private List<CategoryVo> buildTree(List<CategoryVo> categoryList) {
+        Map<String, CategoryVo> categoryMap = new HashMap<>();
+        List<CategoryVo> rootCategories = new ArrayList<>();
+
+        // 首先，使用ID创建一个分类映射
+        for (CategoryVo category : categoryList) {
+            categoryMap.put(category.getId(), category);
+        }
+        // 遍历分类列表，构建树形结构
+        for (CategoryVo category : categoryList) {
+            String parentId = category.getParentId();
+            if (parentId.equals(SysConstants.CATEGORY_PARENT_ID)) {
+                // 顶级分类（没有父分类）
+                rootCategories.add(category);
+            } else {
+                // 子分类
+                CategoryVo parentCategory = categoryMap.get(parentId);
+                if (parentCategory != null) {
+                    if (parentCategory.getChildren() == null) {
+                        parentCategory.setChildren(new ArrayList<>());
+                    }
+                    parentCategory.getChildren().add(category);
+                }
+            }
+        }
+
+        return rootCategories;
+    }
+
+
+    /**
+     * @author: 九思.
+     * @date: 2023/11/17 20:07
      * @param: CategoryVo categoryVo
      * @return: void
      * @description: 新增分类信息
@@ -115,7 +188,7 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
 
         // 如果缩略名为空，自动匹配（纯英文，去掉空格，单词之间下划线分割）
         if (StringUtils.isBlank(shortName)) {
-            shortName = convertShortName(name);
+            categoryVo.setShortName(convertShortName(name));
         } else {
             // 自定义缩略名时重复判断
             // 判断缩略名是否重复
@@ -128,10 +201,9 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
             }
         }
         // 插入标签信息
-        Category newCategory = new Category();
-        newCategory.setName(name);
-        newCategory.setShortName(shortName);
-        categoryMapper.insert(fillDefault(newCategory));
+        Category dbCategory = new Category();
+        BeanUtils.copyProperties(categoryVo, dbCategory);
+        categoryMapper.insert(fillDefault(dbCategory));
     }
 
     /**
@@ -184,8 +256,15 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
      */
     @Override
     public void removeCategorys(List<Long> ids) {
-        // TODO 分类下有文章内容时删除失败
+        // TODO: 分类下有文章内容时删除失败
         ids.forEach(id -> {
+            // 分类下有子分类时，不允许删除
+            LambdaQueryWrapper<Category> wrapper = new LambdaQueryWrapper<>();
+            wrapper.eq(Category::getParentId,id);
+            Integer count = categoryMapper.selectCount(wrapper);
+            if(count > 0){
+                throw new GlobalException(ResultCodeEnum.CATEGORY_DELETE_FAILED);
+            }
             categoryMapper.deleteById(id);
         });
     }
@@ -236,11 +315,11 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
     private Category fillDefault(Category category) {
         // 默认父级id
         if (category.getParentId() == null) {
-            category.setParentId(-1L);
+            category.setParentId(SysConstants.CATEGORY_PARENT_ID);
         }
         // 默认order
         if (category.getOrder() == null) {
-            category.setOrder(5);
+            category.setOrder(SysConstants.ORDER_DEFAULT);
         }
         // 默认逻辑删除标记
         if (StringUtils.isBlank(category.getDelFlag())) {
